Added a mechanism to have hooks be called for proxy methods
In case one function proxies to another function, we have the hooks of
the original function be called.

Although, the problem in this case is that the proxy function may pass
different arguments to the original function and may also change the
return value before returning it.

So, the mechanism may not be safe!