Retry empty text extraction
When text is null for a UI interaction, we might be dealing with
interaction with a sub-elements of a larger sub-tree that represent a
compount component.
In these case, we can try going up the tree and detecting the
interactive root of the sub-tree and then extract the text of that
sub-tree.